feat(management): add health-check information in apis
Closes gravitee-io/issues#4750
